+++ source/log-viewer/index.ts
@@ -17,6 +17,7 @@
 import windowRegister from '../common/modules/window-register'
 import Vue from 'vue'
 import LogViewer from './log-viewer.vue'
+import { ipcRenderer } from 'electron'
 
 // Create the Vue app because we need to reference it in our toolbar controls
 const app = new Vue(LogViewer)
@@ -78,6 +79,13 @@ windowRegister({
       }
     }
   ]
+})
+
+// This window will be closed immediately on a window-close command
+ipcRenderer.on('shortcut', (event, shortcut) => {
+  if (shortcut === 'close-window') {
+    ipcRenderer.send('window-controls', { command: 'win-close' })
+  }
 })
 
 // In the end: mount the app onto the DOM